requirements.txt
requests
python-dotenv
pytz
pyyaml
loguru
PySide6
pyinstaller